4.21*: bloglist expansion, step9

+++ part4/blog-list-app/controllers/blogs.js
@@ -50,6 +50,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token invalid' })
+  }
+
+const blog = await Blog.findById(request.params.id)
+const userId= blog.user.toString()
+	if (decodedToken.id !== userId) {
+		return response.status(401).json({ error: 'invalid user' })
+	}
 	await Blog.findByIdAndDelete(request.params.id)
 	response.status(204).end()
 })
